chore(alloy): cleanup extra data encodings
