+++ crates/alloy/consensus/src/lib.rs
@@ -31,11 +31,8 @@ pub use transaction::{
     OpTransactionInfo, OpTxEnvelope, OpTxType, OpTypedTransaction, TxDeposit,
 };
 
-mod eip1559;
-pub use eip1559::{
-    EIP1559ParamError, decode_eip_1559_params, decode_holocene_extra_data,
-    decode_jovian_extra_data, encode_holocene_extra_data, encode_jovian_extra_data,
-};
+mod extra;
+pub use extra::{EIP1559ParamError, HoloceneExtraData, JovianExtraData};
 
 mod source;
 pub use source::*;
